Abstracting Network and allowing for process spawning via console.
Network class is no longer async only (possibly with strange hacks to make it sync), but it is an abstract class. Two concrete subclasses (SyncNetwork and AsyncNetwork) can be instantiated, depending on needs.

Following the requirements in assignment 2, I introduced the "new" command to spawn new processes locally via the console (instead of creating a new, ideally remote, VM).
Moreover, the assignment 2 requires that IDs can be chosen; this motivated changes in the VectorClock class (now backed by a "sparse" HashMap instead of an array).

The new "local" processes also called for a better handling of clocks: now messages hold a copy of the clock, instead of relying on RMI automatic serialization.