Further subclassing networks.
Separating the classes handling common operations (either sync or async) from the ones specifically related to an assignment. Currently this is a redundand operation, since we are working only on two assignments, but clarifies the code and helps getting ready for the third assignment.

Please note that this commit is not perfeclty working, since there is a half-baked change in SyncNetwork class pertaining to the future works in synchronous communications.